Dockerized angular devenv
Provides a simple dockerized angular development environement.

The idea is to have everything necessary to create and develop an angular application
without installing any of it locally.

Instead npm, node, angular cli, etc. are installed and run in a docker container.

Additionnaly, the current directory is a mounted volume of the docker container to allow the container to
 - write files (e.g. needed to install dependencies required for the app)
 - read and see file changes done locally (e.g. needed to watch files changes with ng serve).